Former-commit-id: 12db28a28f1e86ff508ca9f3ac6fbe45f4437cf7

+++ app/Http/Controllers/Dev/TesteController.php
@@ -338,20 +338,42 @@ class TesteController extends Controller
     public function tgFunction()
     {
         //nada
+//dd('xxx');
+//        $notazInvoiceModel = new NotazzInvoice();
+//        $nservice          = new NotazzService();
+//
+//        $ret = $nservice->consultNfse(1581);
+//        dd($ret);
 
+
+/*----------------------------------------------------------------------------
         $notazInvoiceModel = new NotazzInvoice();
         $nservice          = new NotazzService();
 
-        $invoices = $notazInvoiceModel->get();
+        $invoices = $notazInvoiceModel->where('status','!=',5)->get();
+
+        //        $ret = $nservice->consultNfse(2622);
+        //        dd($ret);
 
         try {
+            $count = 0;
             foreach ($invoices as $invoice) {
+                if($count > 90)
+                {
+                    break;
+                }
                 $ret = $nservice->deleteNfse($invoice->id);
+                if($ret == false)
+                {
+                    continue;
+                }
 
                 $invoice->update([
-                                           'return_message'   => $ret->statusProcessamento,
-                                           'return_http_code' => $ret->codigoProcessamento,
-                                       ]);
+                                     'status'   => 5,
+                                     'return_message'   => $ret->statusProcessamento,
+                                     'return_http_code' => $ret->codigoProcessamento,
+                                 ]);
+                $count = $count + 1;
             }
 
             dd('ok');
@@ -361,6 +383,12 @@ class TesteController extends Controller
         }
 
         dd('fim');
+
+        -------------------------------------------------------------------------------*/
+
+
+
+
         //$nservice->createInvoice(4,3533,1 );
         //dd($nservice->consultNfse('1afe5fa65bf25a023df701e99a64962f'));
